KrowBarApp/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref @ file:///Users/ec2-user/ci_py311/asgiref_1678390266874/work
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
Django @ file:///Users/ec2-user/ci_py311/django_1678411010555/work
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
langchain==0.0.190
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.8
pyspark==3.1.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.15
sqlparse @ file:///Users/ec2-user/ci_py311/sqlparse_1678394278812/work
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
